A porta do tipo "int", passada como argumento para as threads, estava gerando um warning no compilador. O tipo foi alterado para "intptr_t", o que resolveu. Os tamanhos dos envios e recebimentos (funções send e recv) foram alterados todos para 999, pois algumas mensagens estavam falhando.

+++ p2p.c
@@ -18,7 +18,7 @@ int main(int argc, char** argv) {
     pthread_t thread_cliente;
     int fim_servidor, fim_cliente;
     
-    int porta = servidor();
+    intptr_t porta = servidor();
     
     printf("\n\n\n\n\n\n\n\n\n\n");
     printf("\n== P2P =====================================\n");
